requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
transformers==4.35.2
torch>=2.2.0
pydantic==2.4.2
fugashi==1.3.0
unidic-lite==1.0.8